Improved casting of int to int32

+++ blond/_core/backends/cpp/callables.py
@@ -439,14 +439,14 @@ def reload_cpp_backend(
 
         @staticmethod
         def move_flagged_elements_to_end(
-            flag: np.int32,
+            flag: int,
             flags: NumpyArray | CupyArray,  # also purged
             dt: NumpyArray | CupyArray,
             dE: NumpyArray | CupyArray,
             ids: NumpyArray | CupyArray,
         ):
             n_new = _LIBBLOND.move_flagged_elements_to_end(
-                ct.c_int32(flag),
+                ct.c_int32(np.int32(flag)),
                 flags.ctypes.data_as(ct.c_void_p),
                 dt.ctypes.data_as(ct.c_void_p),
                 dE.ctypes.data_as(ct.c_void_p),
